refactor: lazy load docs component
